refactoring finished
