+++ server/routes/userRoutes.ts
@@ -3,6 +3,8 @@ import { userRegister } from "../controllers/user/register.ts";
 import { userLogin } from "../controllers/user/login.ts";
 import { updateProfile } from "../controllers/user/updateProfile.ts";
 import { getUserById } from "../controllers/user/getprofile.ts";
+import { getUsers } from "../controllers/user/getusers.ts";
+
 import { verifyTokenMiddleware } from "../middlewares/middleware.ts";
 
 const router = Router();
@@ -11,5 +13,6 @@ router.post("/register", userRegister);
 router.post("/login", userLogin);
 router.put("/updateprofile", updateProfile);
 router.get("/getprofile/:userId", getUserById);
+router.get("/getusers", verifyTokenMiddleware, getUsers);
 
 export default router;
